% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHpltStats_func_20170924.R
\name{GetHpltStats}
\alias{GetHpltStats}
\title{GetHpltStats() function}
\usage{
GetHpltStats(filepath)
}
\arguments{
\item{filepath}{is a user defined path to the folder where the output files
from the HpltFind() function have been saved.}
}
\value{
A mean of the mean proportion of incongruent sequences for each
  nest.
}
\description{
\code{\link{GetHpltStats}} uses the output files produced by the HpltFind()
function to calculate the mean of the mean proportion of incongruent
sequences across all nests in the data set.
}
\examples{
filepath <- system.file("extdata/HpltFindOut/", package="MHCtools")
GetHpltStats(filepath)
}
\seealso{
\code{\link{HpltFind}}; \code{\link{GetHpltTable}}
}
