% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReplTable_func_20170924.R
\name{GetReplTable}
\alias{GetReplTable}
\title{GetReplTable function}
\usage{
GetReplTable(filepath)
}
\arguments{
\item{filepath}{is a user defined path to the folder where the output files
from the ReplMatch() function have been saved.}
}
\value{
A table with the mean proportion of incongruent sequences for each
  replicate set.
}
\description{
\code{\link{GetReplTable}} uses the output files produced by the ReplMatch()
function to produce a table with the replicate sets and their respective mean
proportion of incongruent sequences.
}
\examples{
filepath <- system.file("extdata/ReplMatchOut/", package="MHCtools")
GetReplTable(filepath)
}
\seealso{
\code{\link{ReplMatch}}; \code{\link{GetReplStats}}
}
