% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplt_occurrence_matrix.R
\docType{data}
\name{hplt_occurrence_matrix}
\alias{hplt_occurrence_matrix}
\title{Data hplt_occurrence_matrix}
\format{
hplt_occurrence_matrix is a data frame with 136 putative haplotypes
  in rows and 329 MHC sequence variants in columns.
}
\source{
original data.
}
\usage{
hplt_occurrence_matrix
}
\description{
hplt_occurrence_matrix is an example of a binary occurrence matrix derived
from a randomized real major histocompatibility complex (MHC) data set. The
matrix was generated using the CreateHpltOccTable() function.
}
\keyword{datasets}
