% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_summary_table.R
\docType{data}
\name{k_summary_table}
\alias{k_summary_table}
\title{k_summary_table.rda}
\format{
k_summary_table is a data frame with observations from
  10 k-estimation scans in rows and their respective stats in 11 columns.
}
\source{
original data.
}
\usage{
k_summary_table
}
\description{
k_summary_table contains the results from a bootstrapped kmeans
clustering analysis performed on the test data in the tables
z1_matrix_test_data, z2_matrix_test_data, z3_matrix_test_data,
z4_matrix_test_data, and z5_matrix_test_data using BootKmeans().
}
\keyword{datasets}
