% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_table.R
\docType{data}
\name{nest_table}
\alias{nest_table}
\title{Data nest_table}
\format{
nest_table is a data frame with 213 samples in rows and 2 columns:
  \describe{ \item{Sample_ID}{Sample ID} \item{Nest}{Nest index number} }
}
\source{
original data.
}
\usage{
nest_table
}
\description{
nest_table, parents_table, and sequence_table comprise a randomized dataset
derived from a real major histocompatibility complex (MHC) genotyping
experiment with data from parents and offspring. The nucleotide sequences
have been replaced with randomly generated sequences, and sample names have
been anonymized.
}
\keyword{datasets}
