% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents_table.R
\docType{data}
\name{parents_table}
\alias{parents_table}
\title{Data parents_table}
\format{parents_table is a data frame with 57 parent pairs in rows and 2
  columns: \describe{ \item{Mother}{Mother ID} \item{Father}{Father ID} }}
\source{
original data.
}
\usage{
parents_table
}
\description{
nest_table, parents_table, and sequence_table comprise a randomized dataset
derived from a real major histocompatibility complex (MHC) genotyping
experiment with data from parents and offspring. The nucleotide sequences
have been replaced with randomly generated sequences, and sample names have
been anonymized.
}
\keyword{datasets}
