% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates_table.R
\docType{data}
\name{replicates_table}
\alias{replicates_table}
\title{Data replicates_table}
\format{
replicates_table is a data frame with 111 technical replicate samples
  in rows and 2 columns: \describe{ \item{Sample_ID}{Technical replicate
  sample ID} \item{Replic_set}{Index number of replicate set} }
}
\source{
original data.
}
\usage{
replicates_table
}
\description{
replicates_table and sequence_table_repl comprise a randomized dataset
derived from a real major histocompatibility complex (MHC) genotyping
experiment with technical replicates. The nucleotide sequences have been
replaced with randomly generated sequences, and sample names have been
anonymized.
}
\keyword{datasets}
