% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table.R
\docType{data}
\name{sequence_table}
\alias{sequence_table}
\title{Data sequence_table}
\format{
sequence_table is a data frame with 334 samples in rows and 329 DNA
  sequence variants in columns.
}
\source{
original data.
}
\usage{
sequence_table
}
\description{
nest_table, parents_table, and sequence_table comprise a randomized dataset
derived from a real major histocompatibility complex (MHC) genotyping
experiment with data from parents and offspring. The nucleotide sequences
have been replaced with randomly generated sequences, and sample names have
been anonymized.
}
\keyword{datasets}
