% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table_fas.R
\docType{data}
\name{sequence_table_fas}
\alias{sequence_table_fas}
\title{Data sequence_table_fas}
\format{sequence_table_fas is a data frame with 100 samples in rows and 166
  DNA sequence variants in columns.}
\source{
original data.
}
\usage{
sequence_table_fas
}
\description{
sequence_table_fas is a randomized dataset derived from a real major
histocompatibility complex (MHC) genotyping experiment. The nucleotide
sequences have been replaced with randomly generated sequences, and sample
names have been anonymized.
}
\keyword{datasets}
