% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table_hplt.R
\docType{data}
\name{sequence_table_hplt}
\alias{sequence_table_hplt}
\title{Data sequence_table_hplt}
\format{
sequence_table_hplt is a data frame with 334 samples in rows and 329
  sequence variants in columns.
}
\source{
original data.
}
\usage{
sequence_table_hplt
}
\description{
sequence_table_hplt is a randomized test data set derived from a real major
histocompatibility complex (MHC) genotyping experiment. This table differs
from the sequence_table by having the nucleotide sequences replaced with
sequence names. Sample names have been anonymized from the real data set.
}
\keyword{datasets}
