% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table_repl.R
\docType{data}
\name{sequence_table_repl}
\alias{sequence_table_repl}
\title{Data sequence_table_repl}
\format{sequence_table_repl is a data frame with 412 samples in rows and 511
  DNA sequence variants in columns.}
\source{
original data.
}
\usage{
sequence_table_repl
}
\description{
replicates_table and sequence_table_repl comprise a randomized dataset
derived from a real major histocompatibility complex (MHC) genotyping
experiment with technical replicates. The nucleotide sequences have been
replaced with randomly generated sequences, and sample names have been
anonymized.
}
\keyword{datasets}
