% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table_SynDist.R
\docType{data}
\name{sequence_table_SynDist}
\alias{sequence_table_SynDist}
\title{Data sequence_table_SynDist}
\format{
sequence_table_SynDist is a data frame with 20 samples in rows
and 20 DNA sequence variants in columns.
}
\source{
original data.
}
\usage{
sequence_table_SynDist
}
\description{
Sequence_table_SynDist comprises an anonymized dataset derived from a real
major histocompatibility complex (MHC) genotyping experiment.
}
\keyword{datasets}
