% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z1_matrix.R
\docType{data}
\name{z1_matrix}
\alias{z1_matrix}
\title{z1_matrix.rda}
\format{
z1_matrix is a data frame with 70 sequences in rows and
  z1-descriptor variables for 8 sequence codons in columns.
}
\source{
original data.
}
\usage{
z1_matrix
}
\description{
z1_matrix comprise a randomized dataset derived from 70
nucleotide sequences of a real major histocompatibility complex (MHC)
genotyping experiment. z1-descriptor values have been extracted from a
subset of 8 amino acid codons and sequence names have been anonymized.
}
\keyword{datasets}
