% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2_matrix.R
\docType{data}
\name{z2_matrix}
\alias{z2_matrix}
\title{z2_matrix.rda}
\format{
z2_matrix is a data frame with 70 sequences in rows and
  z2-descriptor variables for 8 sequence codons in columns.
}
\source{
original data.
}
\usage{
z2_matrix
}
\description{
z2_matrix comprise a randomized dataset derived from 70
nucleotide sequences of a real major histocompatibility complex (MHC)
genotyping experiment. z2-descriptor values have been extracted from a
subset of 8 amino acid codons and sequence names have been anonymized.
}
\keyword{datasets}
