% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z3_matrix.R
\docType{data}
\name{z3_matrix}
\alias{z3_matrix}
\title{z3_matrix.rda}
\format{
z3_matrix is a data frame with 70 sequences in rows and
  z3-descriptor variables for 8 sequence codons in columns.
}
\source{
original data.
}
\usage{
z3_matrix
}
\description{
z3_matrix comprise a randomized dataset derived from 70
nucleotide sequences of a real major histocompatibility complex (MHC)
genotyping experiment. z3-descriptor values have been extracted from a
subset of 8 amino acid codons and sequence names have been anonymized.
}
\keyword{datasets}
