% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z4_matrix.R
\docType{data}
\name{z4_matrix}
\alias{z4_matrix}
\title{z4_matrix.rda}
\format{
z4_matrix is a data frame with 70 sequences in rows and
  z4-descriptor variables for 8 sequence codons in columns.
}
\source{
original data.
}
\usage{
z4_matrix
}
\description{
z4_matrix comprise a randomized dataset derived from 70
nucleotide sequences of a real major histocompatibility complex (MHC)
genotyping experiment. z4-descriptor values have been extracted from a
subset of 8 amino acid codons and sequence names have been anonymized.
}
\keyword{datasets}
