% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z5_matrix.R
\docType{data}
\name{z5_matrix}
\alias{z5_matrix}
\title{z5_matrix.rda}
\format{
z5_matrix is a data frame with 70 sequences in rows and
  z5-descriptor variables for 8 sequence codons in columns.
}
\source{
original data.
}
\usage{
z5_matrix
}
\description{
z5_matrix comprise a randomized dataset derived from 70
nucleotide sequences of a real major histocompatibility complex (MHC)
genotyping experiment. z5-descriptor values have been extracted from a
subset of 8 amino acid codons and sequence names have been anonymized.
}
\keyword{datasets}
