% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFas_func_20170924.R
\name{CreateFas}
\alias{CreateFas}
\title{CreateFas() function}
\usage{
CreateFas(seq_table, path_out)
}
\arguments{
\item{seq_table}{seq_table is a sequence table as output by the 'dada2'
pipeline, which has samples in rows and nucleotide sequence variants in
columns.}

\item{path_out}{is a user defined path to the folder where the output files
will be saved.}
}
\value{
A FASTA file with all the sequences in a 'dada2' sequence table. The
  sequences are named in the FASTA file by an index number corresponding to
  their column number in the sequence table.
}
\description{
\code{\link{CreateFas}} creates a FASTA file with all the sequences in a
'dada2' sequence table.
}
\examples{
seq_table <- sequence_table_fas
path_out <- tempdir()
CreateFas(seq_table, path_out)
}
\seealso{
\code{\link{CreateSamplesFas}}; for more information about 'dada2'
  visit <https://benjjneb.github.io/dada2>
}
